parse_service/requirements.txt
flask
python-dotenv
pdfminer.six
scikit-learn
nltk
python-docx
